requirements.txt
streamlit
python-dotenv
langchain
langchain_huggingface
langsmith
requests
selenium
webdriver-manager
pandas
openpyxl
langchain_community
